--- assets/src/admin/media-sharing/components/syncIcon.tsx
@@ -1,20 +1,0 @@
-/**
- * WordPress dependencies
- */
-import { SVG, Path } from '@wordpress/primitives';
-
-const syncIcon = (
-	<SVG
-		viewBox="0 0 24 24"
-		xmlns="http://www.w3.org/2000/svg"
-		width="24"
-		height="24"
-	>
-		<Path
-			fill="#fff"
-			d="M12 4a8 8 0 017.32 4.91.75.75 0 11-1.38.58A6.5 6.5 0 0012 5.5c-1.73 0-3.29.69-4.42 1.8l1.57.01a.75.75 0 110 1.5H5.25a.75.75 0 01-.75-.75V5.5a.75.75 0 111.5 0v1.66A7.98 7.98 0 0112 4zm0 16a8 8 0 01-7.32-4.91.75.75 0 011.38-.58A6.5 6.5 0 0012 18.5c1.73 0 3.29-.69 4.42-1.8l-1.57-.01a.75.75 0 110-1.5h3.9a.75.75 0 01.75.75v3.3a.75.75 0 11-1.5 0v-1.66A7.98 7.98 0 0112 20z"
-		/>
-	</SVG>
-);
-
-export default syncIcon;
