--- assets/src/admin/media-sharing/components/versionIcon.tsx
@@ -1,21 +1,0 @@
-/**
- * WordPress dependencies
- */
-import { SVG, Path } from '@wordpress/primitives';
-
-const versionIcon = (
-	<SVG
-		viewBox="0 0 510 510"
-		xmlns="http://www.w3.org/2000/svg"
-		width="24"
-		height="24"
-	>
-		<Path
-			fill="#000"
-			d="M0 0 C0.6886783 -0.00265183 1.37735659 -0.00530365 2.08690393 -0.00803584 C4.4056911 -0.01572462 6.72442527 -0.01628785 9.04322433 -0.01688576 C10.71819306 -0.02090069 12.39316081 -0.02533897 14.06812739 -0.03016531 C18.68004159 -0.04213382 23.29194533 -0.04777994 27.90387201 -0.05234647 C32.87484963 -0.05830174 37.84581473 -0.069912 42.81678391 -0.08063698 C54.82864462 -0.10516619 66.84050949 -0.11979069 78.85238755 -0.13269705 C82.24599905 -0.13640917 85.63961015 -0.14038241 89.03322124 -0.14444971 C110.11211934 -0.16969093 131.1910181 -0.19174173 152.26992798 -0.20417786 C157.15324277 -0.20710752 162.03655754 -0.21006918 166.91987228 -0.21308327 C168.13358414 -0.21382932 169.34729599 -0.21457537 170.59778699 -0.21534403 C190.24500842 -0.22798885 209.89213917 -0.26119948 229.53931823 -0.3028851 C249.71702415 -0.34534473 269.89468372 -0.36887772 290.07243419 -0.37409842 C301.39877069 -0.3775084 312.72496561 -0.38921462 324.05125809 -0.42220116 C333.7008564 -0.45025844 343.35028411 -0.46117618 352.99992004 -0.4500057 C357.91941371 -0.44483321 362.83855041 -0.44761374 367.75799179 -0.47247124 C372.2714339 -0.49508572 376.78426516 -0.49413001 381.29771885 -0.47526737 C382.92103108 -0.47248431 384.54438582 -0.47797039 386.16763011 -0.4930803 C405.80819551 -0.66456235 421.3792 5.14177897 435.89936447 18.76592064 C449.61956272 33.02646916 453.48824249 49.40918497 453.44804001 68.5008297 C453.45177882 70.18138976 453.45638126 71.86194808 453.46177065 73.54250365 C453.4739876 78.15119301 453.47374172 82.75980971 453.4713819 87.36851156 C453.4709734 92.34277421 453.48201607 97.31701367 453.49149704 102.29126549 C453.50810108 112.03355533 453.51360633 121.77582096 453.5147218 131.51812377 C453.51567259 139.43878279 453.51978448 147.35943383 453.52612305 155.28009033 C453.54373418 177.74468716 453.55296217 200.20926489 453.5514621 222.67386868 C453.55134223 224.49000203 453.55134223 224.49000203 453.55121994 226.34282494 C453.55113818 227.55506172 453.55105642 228.76729849 453.55097218 230.01626972 C453.55015941 249.65878564 453.56929502 269.30122484 453.59749667 288.94371857 C453.62623631 309.1204115 453.64003636 329.29706702 453.63830912 349.47378093 C453.63764557 360.79825665 453.64312246 372.12265196 453.66466522 383.44710922 C453.68289124 393.08973413 453.68721267 402.73224219 453.67364638 412.37487688 C453.66712382 417.2919468 453.66697381 422.20879264 453.68435287 427.12584114 C453.70012017 431.6333525 453.69718273 436.14046278 453.67978461 440.64796457 C453.67672275 442.27244309 453.68031622 443.89695182 453.69150476 445.52139469 C453.81853214 465.28433045 447.70713147 480.81790782 434.19233322 495.26201439 C429.85006038 499.448914 425.12267488 502.55372126 419.86811447 505.44951439 C419.28860031 505.7751638 418.70908615 506.10081322 418.11201096 506.4363308 C407.10458881 512.19534938 395.52868867 512.57284462 383.40497971 512.51796532 C381.72850332 512.52142749 380.05202921 512.5261958 378.37555981 512.53214705 C373.79673766 512.54460978 369.21810436 512.5385083 364.63928175 512.52918839 C359.69032896 512.52237879 354.74140782 512.53288859 349.7924614 512.54112816 C340.10852172 512.55444732 330.42465192 512.55159585 320.74070954 512.54289024 C312.86828556 512.53610695 304.99588112 512.53519976 297.12345505 512.53846169 C295.44089588 512.53915123 295.44089588 512.53915123 293.72434562 512.53985471 C291.4454108 512.54081392 289.16647599 512.54178656 286.88754117 512.54277249 C265.53279461 512.55132106 244.17808672 512.5415006 222.8233465 512.52537139 C204.5117524 512.5119687 186.20021535 512.51429694 167.88862228 512.52812767 C146.6087427 512.54419288 125.32889699 512.55049713 104.04901218 512.54129088 C101.77859943 512.54033526 99.50818666 512.53939215 97.2377739 512.53846169 C96.12077352 512.53799886 95.00377315 512.53753603 93.85292429 512.53705918 C85.99265346 512.53447212 78.13240418 512.53883417 70.27213669 512.54590034 C60.68924289 512.55432599 51.10642279 512.55201279 41.5235374 512.53605811 C36.63698316 512.52818977 31.75055166 512.5250593 26.86399841 512.53500557 C22.38574014 512.54398273 17.90771773 512.53897007 13.42948067 512.52307291 C11.81456487 512.51973787 10.19962891 512.52139024 8.58472621 512.52870028 C-10.56114267 512.60867871 -26.7115085 507.75205822 -40.80791092 494.41631126 C-53.36851101 481.83629757 -59.0582388 465.9205484 -59.09619653 448.48634034 C-59.09325472 446.9152929 -59.09011932 445.3442458 -59.08681107 443.77319908 C-59.09054988 442.09263902 -59.09515232 440.4120807 -59.10054171 438.73152512 C-59.11275866 434.12283577 -59.11251278 429.51421907 -59.11015296 424.90551722 C-59.10974446 419.93125457 -59.12078713 414.95701511 -59.1302681 409.98276329 C-59.14687214 400.24047345 -59.15237738 390.49820782 -59.15349285 380.75590501 C-59.15444365 372.83524599 -59.15855554 364.91459494 -59.1648941 356.99393845 C-59.18250523 334.52934161 -59.19173323 312.06476388 -59.19023316 289.6001601 C-59.19015324 288.38940453 -59.19007333 287.17864897 -59.189991 285.93120384 C-59.18990924 284.71896706 -59.18982747 283.50673028 -59.18974324 282.25775906 C-59.18893047 262.61524314 -59.20806608 242.97280393 -59.23626773 223.3303102 C-59.26500737 203.15361728 -59.27880741 182.97696176 -59.27708018 162.80024785 C-59.27641663 151.47577213 -59.28189351 140.15137682 -59.30343628 128.82691956 C-59.3216623 119.18429465 -59.32598373 109.54178659 -59.31241743 99.8991519 C-59.30589488 94.98208197 -59.30574487 90.06523614 -59.32312393 85.14818764 C-59.33889122 80.64067628 -59.33595379 76.133566 -59.31855567 71.6260642 C-59.3154938 70.00158569 -59.31908728 68.37707696 -59.33027582 66.75263409 C-59.45659936 47.09920152 -53.45019515 31.13061991 -39.69047928 16.94170189 C-29.12570493 7.00060832 -14.6178972 0.02771614 0 0 Z M173.18061447 68.13701439 C172.28004318 68.35132103 171.37947189 68.56562767 170.45161057 68.78642845 C126.10799004 79.69176846 86.05242066 106.15662826 59.18061447 143.13701439 C58.76215256 143.7111306 58.34369064 144.28524681 57.91254807 144.87676048 C46.15703777 161.22806528 37.56145836 179.08159322 31.18061447 198.13701439 C30.67272385 197.42416283 30.16483322 196.71131126 29.64155197 195.97685814 C21.48920923 183.45962435 21.48920923 183.45962435 8.89545822 176.60185814 C2.44798538 175.57544438 -2.37608051 176.47345154 -7.81938553 180.13701439 C-12.92210982 184.85730753 -16.08783265 189.5797453 -16.38188553 196.63701439 C-16.24609033 201.54626388 -15.36924688 204.58137121 -12.80376053 208.75810814 C-12.16639969 209.80015404 -11.52903885 210.84219994 -10.87236404 211.91582298 C-7.4878711 217.22620471 -4.02860499 222.47918388 -0.50688553 227.69951439 C1.07876575 230.05690508 2.66282235 232.41536593 4.24702072 234.77373314 C4.6505275 235.37352989 5.05403427 235.97332664 5.46976852 236.59129906 C8.39490254 240.94582009 11.27703149 245.32702128 14.14545822 249.71904564 C15.35137901 251.55732796 16.55819791 253.39501675 17.76542282 255.23244286 C18.53448539 256.40701517 19.30036562 257.58367811 20.06284714 258.76253319 C25.22822086 266.7435392 29.68535666 273.59461922 39.04780197 276.69170189 C47.81903771 278.5230588 54.40161604 274.17507121 61.54032516 269.63692284 C64.03226575 267.99698438 66.50542868 266.33206071 68.97358322 264.65654564 C69.8820752 264.04365181 70.79056717 263.43075798 71.7265892 262.79929161 C73.63961078 261.50760853 75.55143341 260.21414815 77.46210861 258.91899681 C80.34004043 256.96854096 83.22189218 255.02399552 86.1044426 253.08037376 C91.47702667 249.45272459 96.83453937 245.80360753 102.18061447 242.13701439 C102.92297348 241.62878136 103.66533249 241.12054832 104.43018723 240.59691429 C106.39139752 239.24842849 108.34434267 237.88899912 110.29389572 236.52373314 C111.35004074 235.78727562 112.40618576 235.0508181 113.49433517 234.29204369 C118.13180068 230.57170305 120.78970386 226.93247526 122.18061447 221.13701439 C122.69747309 215.55494129 121.78869489 211.05303898 119.18061447 206.13701439 C116.84712247 203.30290284 116.84712247 203.30290284 114.18061447 201.13701439 C113.56186447 200.58013939 112.94311447 200.02326439 112.30561447 199.44951439 C107.49407619 196.47768192 102.77571578 195.67852508 97.20038986 196.59941673 C90.58223642 198.36095111 84.76588458 202.56726704 79.18061447 206.38701439 C78.50127853 206.84592064 77.8219426 207.30482689 77.12202072 207.77763939 C75.47198663 208.89315539 73.8258736 210.0144678 72.18061447 211.13701439 C76.60415561 190.21735108 88.20426398 171.03943439 102.18061447 155.13701439 C102.91022385 154.28623314 103.63983322 153.43545189 104.39155197 152.55888939 C129.53300027 123.80060116 165.21943167 107.58912982 202.86811447 103.38701439 C246.80111635 100.58132424 287.05506592 114.27375922 320.09077072 143.30498314 C344.93861005 165.75243911 361.28563433 198.15237898 366.18061447 231.13701439 C366.27342697 231.75793968 366.36623947 232.37886497 366.46186447 233.01860619 C368.95784884 252.03488708 368.50256361 273.44438539 364.18061447 292.13701439 C363.90652756 293.32448265 363.90652756 293.32448265 363.62690353 294.53594017 C359.09304154 313.20368887 350.82114349 331.80146267 339.18061447 347.13701439 C338.55928635 347.96588158 337.93795822 348.79474876 337.29780197 349.64873314 C328.98182389 360.37196803 320.02307175 369.94952832 309.18061447 378.13701439 C308.01659103 379.02453392 308.01659103 379.02453392 306.82905197 379.92998314 C273.89379235 403.99805748 232.85221352 413.19798403 192.66303635 407.53496361 C151.90871029 401.00176618 116.87684561 378.11977685 92.63764572 345.11748314 C87.64008288 339.05525667 82.83027649 335.66339122 74.86811447 334.76201439 C68.35078143 335.42479402 63.15289418 337.64846778 58.74311447 342.57451439 C54.23599568 348.49244876 53.31693978 353.77305123 54.18061447 361.13701439 C57.62702792 373.5441028 70.50025057 385.02892242 79.18061447 394.13701439 C79.9463176 394.96717064 80.71202072 395.79732689 81.50092697 396.65263939 C114.75051615 431.59288565 165.06115881 450.49408232 212.64155197 452.10185814 C266.92933138 453.03110842 315.14828129 430.80375162 354.18061447 394.13701439 C355.13387619 393.26496361 355.13387619 393.26496361 356.10639572 392.37529564 C385.98508998 364.06541111 404.78798142 320.52641961 409.18061447 280.13701439 C409.34496994 278.79896751 409.34496994 278.79896751 409.51264572 277.43388939 C414.68738507 228.03237772 399.89408562 177.11328188 369.18061447 138.13701439 C358.13634482 124.55427366 346.24233001 112.57493875 332.18061447 102.13701439 C331.37019699 101.53542503 331.37019699 101.53542503 330.54340744 100.92168236 C286.11072074 68.31216822 227.03533077 55.19770241 173.18061447 68.13701439 Z M197.40717697 156.35576439 C195.12154914 159.6750201 193.93220212 162.26105486 193.91318893 166.3226223 C193.90465897 167.14956093 193.896129 167.97649956 193.88734055 168.82849693 C193.88824692 169.72727058 193.88915329 170.62604424 193.89008713 171.55205345 C193.8833699 172.5053907 193.87665268 173.45872795 193.8697319 174.44095421 C193.85072353 177.59490221 193.84682597 180.74863765 193.84467697 183.90263939 C193.83825 186.09736277 193.83152172 188.29208529 193.82450485 190.48680687 C193.81252426 195.09011889 193.80882474 199.69335584 193.810009 204.29668236 C193.810212 210.18572781 193.78293287 216.07428964 193.74846458 221.9632206 C193.72618707 226.49808002 193.72217176 231.03282332 193.72322845 235.56773186 C193.7208729 237.73818437 193.712052 239.9086401 193.69656372 242.07903862 C193.67689121 245.12116518 193.68274763 248.16225265 193.69477463 251.2043972 C193.68353561 252.09507393 193.6722966 252.98575066 193.66071701 253.90341759 C193.74573636 262.39846291 196.75287394 268.17600471 202.7232933 274.10722923 C203.2931546 274.65330833 203.8630159 275.19938744 204.45014572 275.76201439 C205.04937347 276.35321068 205.64860123 276.94440697 206.2659874 277.5535183 C208.16569499 279.42449367 210.07856339 281.28124232 211.99311447 283.13701439 C214.50381293 285.59109498 217.00909033 288.05064896 219.51264572 290.51201439 C220.11395813 291.10178062 220.71527054 291.69154686 221.33480453 292.2991848 C224.8919572 295.80577185 228.26977172 299.39281504 231.50019455 303.20537376 C235.1852382 307.44132689 239.17189421 311.39638805 243.11811447 315.38701439 C243.97340744 316.27131126 244.82870041 317.15560814 245.70991135 318.06670189 C251.85616587 324.27390325 258.0047868 329.76249081 267.18061447 330.13701439 C274.01757229 329.76007185 278.41308754 327.13388488 283.28998947 322.41045189 C287.23311389 317.66892624 288.48746993 312.74590097 288.24311447 306.63701439 C286.63418357 297.78888072 280.74851837 292.39386307 273.66602707 287.42682457 C257.42158439 276.00932413 238.11910573 261.16503464 234.42448521 240.54062653 C231.42693975 219.31497819 232.78191173 197.13877267 234.63758469 175.85306358 C235.6465869 164.92218221 235.6465869 164.92218221 231.18061447 155.13701439 C227.42511563 150.78500455 222.92665612 147.59895401 217.15327072 146.88701439 C208.45259484 146.41107998 202.80913571 149.7164855 197.40717697 156.35576439 Z"
-			transform="translate(58.81938552856445,-0.13701438903808594)"
-		/>
-	</SVG>
-);
-
-export default versionIcon;
