--- assets/src/components/SiteTable.tsx
@@ -1,175 +1,0 @@
-/**
- * WordPress dependencies
- */
-/**
- * External dependencies
- */
-import { useState } from 'react';
-import {
-	Button,
-	Card,
-	CardHeader,
-	CardBody,
-	Modal,
-} from '@wordpress/components';
-import { __ } from '@wordpress/i18n';
-/**
- * Internal dependencies
- */
-import type { BrandSite, EditingIndex } from '../types/settings';
-
-const SiteTable = ( {
-	sites,
-	onEdit,
-	onDelete,
-	setFormData,
-	setShowModal,
-}: {
-	sites: BrandSite[];
-	onEdit: ( index: number ) => void;
-	onDelete: ( index: number | null ) => void;
-	setFormData: ( data: BrandSite ) => void;
-	setShowModal: ( show: boolean ) => void;
-} ) => {
-	const [ showDeleteModal, setShowDeleteModal ] = useState( false );
-	const [ deleteIndex, setDeleteIndex ] = useState< EditingIndex >( null );
-
-	const handleDeleteClick = ( index: number ) => {
-		setDeleteIndex( index );
-		setShowDeleteModal( true );
-	};
-
-	const handleDeleteConfirm = () => {
-		onDelete( deleteIndex );
-		setShowDeleteModal( false );
-		setDeleteIndex( null );
-	};
-
-	const handleDeleteCancel = () => {
-		setShowDeleteModal( false );
-		setDeleteIndex( null );
-	};
-
-	return (
-		<Card style={ { marginTop: '30px' } }>
-			<CardHeader>
-				<h3>{ __( 'Brand Sites', 'onemedia' ) }</h3>
-				<Button
-					style={ { width: 'fit-content' } }
-					variant="primary"
-					onClick={ () => setShowModal( true ) }
-				>
-					{ __( 'Add Brand Site', 'onemedia' ) }
-				</Button>
-			</CardHeader>
-			<CardBody>
-				<table
-					className="wp-list-table widefat fixed striped "
-					style={ { marginTop: '16px' } }
-				>
-					<thead>
-						<tr>
-							<th>{ __( 'Site Name', 'onemedia' ) }</th>
-							<th>{ __( 'Site URL', 'onemedia' ) }</th>
-							<th>{ __( 'API Key', 'onemedia' ) }</th>
-							<th>{ __( 'Actions', 'onemedia' ) }</th>
-						</tr>
-					</thead>
-					<tbody>
-						{ sites.length === 0 && (
-							<tr>
-								<td
-									colSpan={ 4 }
-									style={ { textAlign: 'center' } }
-								>
-									{ __(
-										'No Brand Sites found.',
-										'onemedia'
-									) }
-								</td>
-							</tr>
-						) }
-						{ sites?.map( ( site, index ) => (
-							<tr key={ index }>
-								<td>{ site?.name }</td>
-								<td>{ site?.url }</td>
-								<td>
-									<code>
-										{ site?.api_key?.substring( 0, 10 ) }...
-									</code>
-								</td>
-								<td>
-									<Button
-										variant="secondary"
-										onClick={ () => {
-											setFormData( site );
-											onEdit( index );
-											setShowModal( true );
-										} }
-										style={ { marginRight: '8px' } }
-									>
-										{ __( 'Edit', 'onemedia' ) }
-									</Button>
-									<Button
-										variant="secondary"
-										isDestructive
-										onClick={ () =>
-											handleDeleteClick( index )
-										}
-									>
-										{ __( 'Delete', 'onemedia' ) }
-									</Button>
-								</td>
-							</tr>
-						) ) }
-					</tbody>
-				</table>
-			</CardBody>
-			{ showDeleteModal && (
-				<DeleteConfirmationModal
-					onConfirm={ handleDeleteConfirm }
-					onCancel={ handleDeleteCancel }
-				/>
-			) }
-		</Card>
-	);
-};
-
-const DeleteConfirmationModal = ( {
-	onConfirm,
-	onCancel,
-}: {
-	onConfirm: () => void;
-	onCancel: () => void;
-} ) => (
-	<Modal
-		title={ __( 'Delete Brand Site', 'onemedia' ) }
-		onRequestClose={ onCancel }
-		isDismissible
-		shouldCloseOnClickOutside
-	>
-		<p>
-			{ __(
-				'Are you sure you want to delete this Brand Site? This action cannot be undone.',
-				'onemedia'
-			) }
-		</p>
-		<div
-			style={ {
-				display: 'flex',
-				justifyContent: 'flex-end',
-				marginTop: '20px',
-				gap: '16px',
-			} }
-		>
-			<Button variant="secondary" onClick={ onCancel }>
-				{ __( 'Cancel', 'onemedia' ) }
-			</Button>
-			<Button variant="primary" isDestructive onClick={ onConfirm }>
-				{ __( 'Delete', 'onemedia' ) }
-			</Button>
-		</div>
-	</Modal>
-);
-
-export default SiteTable;
